+++ frontend/src/metabase/actions/containers/ActionCreator/ActionCreator.tsx
@@ -3,16 +3,16 @@ import type { Route } from "react-router";
 import { t } from "ttag";
 import _ from "underscore";
 
-import { skipToken, useGetActionQuery } from "metabase/api";
+import {
+  skipToken,
+  useCreateActionMutation,
+  useGetActionQuery,
+  useUpdateActionMutation,
+} from "metabase/api";
 import { LeaveRouteConfirmModal } from "metabase/common/components/LeaveConfirmModal";
 import { Modal } from "metabase/common/components/Modal";
 import { useBeforeUnload } from "metabase/common/hooks/use-before-unload";
 import { useCallbackEffect } from "metabase/common/hooks/use-callback-effect";
-import type {
-  CreateActionParams,
-  UpdateActionParams,
-} from "metabase/entities/actions";
-import { Actions } from "metabase/entities/actions";
 import { Databases } from "metabase/entities/databases";
 import { Questions } from "metabase/entities/questions";
 import { connect } from "metabase/redux";
@@ -55,32 +55,17 @@ interface StateProps {
   metadata: Metadata;
 }
 
-interface DispatchProps {
-  onCreateAction: (params: CreateActionParams) => void;
-  onUpdateAction: (params: UpdateActionParams) => void;
-}
-
 export type ActionCreatorProps = OwnProps;
 
-type Props = OwnProps & ModelLoaderProps & StateProps & DispatchProps;
+type Props = OwnProps & ModelLoaderProps & StateProps;
 
 const mapStateToProps = (state: State) => ({
   metadata: getMetadata(state),
 });
 
-const mapDispatchToProps = {
-  onCreateAction: Actions.actions.create,
-  onUpdateAction: Actions.actions.update,
-};
-
-function ActionCreator({
-  model,
-  onCreateAction,
-  onUpdateAction,
-  onSubmit,
-  onClose,
-  route,
-}: Props) {
+function ActionCreator({ model, onSubmit, onClose, route }: Props) {
+  const [createAction] = useCreateActionMutation();
+  const [updateAction] = useUpdateActionMutation();
   const {
     action,
     formSettings,
@@ -112,12 +97,11 @@ function ActionCreator({
       return; // only query action creation is supported now
     }
 
-    const reduxAction = await onCreateAction({
+    const createdAction = await createAction({
       ...action,
       ...values,
       visualization_settings: formSettings,
-    } as WritebackQueryAction);
-    const createdAction = Actions.HACK_getObjectFromAction(reduxAction);
+    } as WritebackQueryAction).unwrap();
 
     // Sync the editor state with data from save modal form
     handleActionChange(values);
@@ -132,13 +116,12 @@ function ActionCreator({
 
   const handleUpdate = async () => {
     if (isSavedAction(action)) {
-      const reduxAction = await onUpdateAction({
+      const updatedAction = await updateAction({
         ...action,
         model_id: model?.id(),
         visualization_settings: formSettings,
-      });
+      }).unwrap();
 
-      const updatedAction = Actions.HACK_getObjectFromAction(reduxAction);
       onSubmit?.(updatedAction);
 
       scheduleCallback(() => {
@@ -231,5 +214,5 @@ export default _.compose(
     entityAlias: "model",
   }),
   Databases.loadList(),
-  connect(mapStateToProps, mapDispatchToProps),
+  connect(mapStateToProps),
 )(ActionCreatorWithContext);
