+++ frontend/src/metabase/actions/containers/ActionCreator/CreateActionForm/CreateActionForm.tsx
@@ -2,13 +2,13 @@ import { useMemo } from "react";
 import { t } from "ttag";
 import * as Yup from "yup";
 
+import type { CreateQueryActionParams } from "metabase/actions/types";
 import { Button } from "metabase/common/components/Button";
 import { FormErrorMessage } from "metabase/common/components/FormErrorMessage";
 import { FormFooter } from "metabase/common/components/FormFooter";
 import { FormInput } from "metabase/common/components/FormInput";
 import { FormSubmitButton } from "metabase/common/components/FormSubmitButton";
 import { FormTextArea } from "metabase/common/components/FormTextArea";
-import type { CreateQueryActionParams } from "metabase/entities/actions";
 import { Form, FormProvider } from "metabase/forms";
 import { FormModelPicker } from "metabase/models/containers/FormModelPicker";
 import * as Errors from "metabase/utils/errors";
